+++ effective-gear/fabric/src/main/java/com/chimericdream/effectivegear/fabric/client/EffectiveGearFabricClient.java
@@ -1,6 +1,7 @@
 package com.chimericdream.effectivegear.fabric.client;
 
 import com.chimericdream.effectivegear.client.Keybindings;
+import com.chimericdream.effectivegear.client.PreservingBlockColors;
 import com.chimericdream.effectivegear.network.UseAbilityPayload;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
@@ -10,6 +11,7 @@ public final class EffectiveGearFabricClient implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
         Keybindings.init();
+        PreservingBlockColors.init();
 
         ClientTickEvents.END_CLIENT_TICK.register(client -> {
             while (Keybindings.USE_ABILITY.consumeClick()) {